Point tsconfig to client and server TS files only
Previous config defined a glob for all TS files. But only the files
required from the client or server root files need to be transpiled.
This prevents the whole project build failinig due to one corrupt source
file that might not be committed to the repository, and isn't actually
used anywhere.